funciona el codigo comentarios

+++ comentarios/src/main/java/comentarios/comentarios/DTO/ComentarioHotelDTO.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class ComentarioHotelDTO {
     private String nombreHotel;
     private int reserva_id; // Ajusta a Long si tus IDs son Long
-    private int puntuacion;
+    private double puntuacion;
     private String comentario;
 }
